Move test_search.py
